Add access restriction on some URL patterns
